+++ src/vidigi/__init__.py
@@ -2,6 +2,7 @@ from . import (
     analysis,
     animation,
     ciw,
+    plots,
     prep,
     utils,
     logging,